Add files via upload
Tried doing this on the desktop app and it change anything. Managed to figure out how to do if from the browser. If there is somehow any duplicates of the sprite folder just delete it as I am not sure why GitHub desktop wasnt wanting to work correctly for me.